refactor google function + add dotenv
