+++ src/app.ts
@@ -2,7 +2,10 @@ import updateNotifier from 'update-notifier';
 import * as commands from './commands';
 import * as commandsBase from './commands-base';
 
+import { config } from 'dotenv';
+
 const pkg = require('../package.json');
+config();
 
 const notifier = updateNotifier({
   pkg,
